promethea-engine copy 2/mcp_servers/discourse_mcp/requirements.txt
fastapi
uvicorn[standard]
praw
python-dotenv
